[WIP] Avoid merging types with errors.
BUG=27626735

Rationale:

Sending this out for discussion, to make sure all implications
are thoroughly worked out. The underlying problem was that we
tried to merge two classes, where one had a link error. As a result,
the error type thought it was Object (mirror's method  IsObjectClass()
returns true if there is no superclass).

This fix avoids the problem at the client where it crashed. However,
does this suffice? Should we deal with this at other places too,
perhaps even in mirror's methods? Or should we never have seen
this error and am I chasing the wrong fix?

Change-Id: I0fe47bcb5b12af8f0a2315f411c3693152abf880