[WIP] Avoid resolving methods that do not match fundamental return type.
Rationale:
Sending this out for discussion, as this explores new territories for me.
This sanity check avoids accepting invoke-super to methods where the
fundamental return type does not match (so L vs L is okay, but L vs V
is not). This avoids building a HIR with inconsistencies that crash.
I suspect the invoke-super is e.g. the result of proguarding very
aggressively, but, again, I am not very familiar with this stuff yet.

Change-Id: I0562a636ff76a1ee176efef4a575fd6cede1df86