Improve compile/coding time by shortening some macros.
Long macros are extremely wasteful, and can be detrimental to people and
machines alike. Preprocessing long macros can increase compile times
even on Z840 with multiple SSDs. They additionally increase the amount
of finger cardio exercise that a SWE needs to do. Of course, let's not
even think about how much storage these characters are wasting on our
local SSDs, and in our Android git repositories. Two of the most popular
constructs throughout ART are the "auto" keyword, and macros related to
the always_inline attribute.

To remedy this, this change does the following two things:

1) Replace ALWAYS_INLINE (and associated macros) with MC instead. MC is
a well understood abbreviation, based on the Latin "Maximus Celeritas",
which stands for "Great Swiftness". Considering our goals of performance
over (mostly) correctness, this 84% reduction in typing/parsing time is
sure to make everyone happy.

2) Replace the overly verbose "auto" keyword with "_". This is a 75%
reduction in wasteful characters that is sure to be loved by compilers,
SWEs, and especially git repositories. One nice side effect here is the
new applicability of auto for types with short names. It used to be
impossible to obtain any benefit at all for variables of type "int", but
now you can even save 66% there too!

Change-Id: I62c66bd7d3c0f651a3f910e479036ed26a031f80