Avoid falling into infinite loop when owner of lock is lost
To fix bug https://code.google.com/p/android/issues/detail?id=206066
If we can't find owner id in thread list in SuspendThreadByThreadId, it will return null and timed_out is false.
Then ThinLock will not Inflate. And it will be thin lock forever and fall into infinite loop.

So I just let it Inflate when it occurs. There are three conditions in if(!timed_out)) condition:
1. We succeed in suspending the thread. Then just do as what it do before we mofify.
2. SuspendThreadByThreadId return null since owner thread unlocks and exits. At this time, lock_word.GetState() != LockWord::kThinLocked or lock_word.ThinLockOwner() != owner_thread_id, so it will not Inflate. And go back to while loop outside.
3. SuspendThreadByThreadId return null since owner thread don't give up lock and owner thread don't exist. This is the problem I found. Then it will Inflate to fat lock using null owner, which works normally.
In Inflate(...), it will do process of multi-threads.

Signed-off-by: ganxiaolin <ganxiaolin@xiaomi.com>
Change-Id: Iba0aa912d7bbbc71f6ce0cab1539bdc848d32efb