ART: Update DCHECKs in SsaLivenessAnalysis::AddBackEdgeUses
Graph linearization in the presence of irreducible loops is not
guaranteed to generate a linear order where all blocks of a loop are
adjacent, first block is the header and last block is one of the back
edges.

These assumptions are made when inserting synthesized uses at the back
edges to aid the register allocator. Not meeting them will result in
the algorithm's early termination and back-edge uses not being added.

This patch updates the DCHECKs so the compiler does not fail in such
circumstances.

Bug: 27615840
Bug: 27624868

Change-Id: I63632e8819ea3644d5c6fdfea00b66128bf22c24