Update `ArmInstructionSetFeatures` to track ARMv8-A.
Some instructions introduced in ARMv8-A will be useful to improve
the generated code.

Note that we remove mentions of cortex R and M cores from supported
variants. Android is only compiled for A class cores. Mentions of
cortex R and M class cores implies that the features detection should
work for them. This places additional constraints on the detection and
may prevent correct detection of some ARMv8-A features.

Change-Id: I09ff7f5e8b41c872efe7d5581596040da7479d60