Generate debug info for non-debuggables
Native debug info is needed not only for --debuggable, but for any other
compilations with -g. --debuggable drastically affects the generated code
(e.g. prevents inlining) and can be inconvenient for performance
investigations.

For accuracy of hardware level performance profiling this patch avoids
generating extra nops if --debuggable is not set. This results in code
without extra nop instructions as if neither -g nor --debuggable is set.

Change-Id: I1afaf275c62d2f72fc3aea22852e87d830dc237c
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>