Fix JIT inliner problem with nested inlining
While playing with the inliner, I ran into this problem.  I don't have a
test case that I can use to reproduce this under AOSP.  It does seem
like a safe check in any case.

Under JIT mode, the inliner looks up the dex PC for the current invoke
to find the matching InlineCache entry.  If the current code comes from
an already inlined routine, the profiling entry will not match with the
current method.

Fix:  If the HIR is from a nested inline (detected through
IsFromInlinedInvoke()), skip that entry.  The assumption is that the
invoke will have been processed when handling the nested inline.

Change-Id: I3430422b69321f847db80b69df7abaa2a7844f7a
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>