Remove a spurious DCHECK in ModifySuspendCount.
To avoid a failure of the DCHECK in the jdwp test.

ThreadList::RunCheckpoint and ThreadList::FlipThreadRoots calls
ModifySuspendCount on a suspended thread without holding the thread list
lock and with the thread != the current thread, which is fine. But
Dbg::Disconnected uses SetStateUnsafe to switch from suspended to
runnable, which causes the jdwp thread to look runnable and the DCHECK
to fail on a concurrent ModifySuspendCount() call.

As the thread list lock doesn't guard any data touched in
ModifySuspendCount(), I believe the DCHECK is safe to remove. See
25262836#9 for details.

Bug: 25262836
Change-Id: I8c89ce94628fc0bb63855a0b094bdf33b33db172