Disable a DCHECK in ModifySuspendCount when thread is suspended.
To avoid a spurious failure of the DCHECK in the jdwp test.

ThreadList::RunCheckpoint and ThreadList::FlipThreadRoots calls
ModifySuspendCount on a suspended thread without holding the thread list
lock and with the thread != the current thread, which is fine. But
Dbg::Disconnected uses SetStateUnsafe to switch from suspended to
runnable, which causes the jdwp thread to look runnable and the DCHECK
to fail on a concurrent ModifySuspendCount() call.

As the DCHECK asserts the thread list lock is held to make sure that the
thread won't exit and the thread is known to be suspended in
RunCheckpoint, we can disable the DCHECK in this particular case.

Bug: 25262836
Change-Id: I8c89ce94628fc0bb63855a0b094bdf33b33db172