Checker: Refactor line-matching algorithm
Refactor Checker's algorithm which matches assertions against
C1visualizer output. Evaluation of DAG and NOT assertions previously
relied on assertions which need to be evaluated in the same scope
to be adjacent in the Checker test source code. This refactoring
lifts this constraint in anticipation of adding control-flow
statements which will break up the DAG/NOT groups.

Change-Id: I3cc53e9f62c7ea5442dc759b066a262e49621b5f