Change IMT size from 64 to 43
Data shows that this change can save around 2MB with an estimated
less than 2% regression in interface method invokation speed.

Further investigation is needed to measure the performance impact
in practice, particularly for much smaller IMT sizes. We are making
this initial change to 43 to see if it affects runtime perf benchmarks.

Change-Id: Ic0e036091c5da0e6b93f519fc695e9b48ab5ec46