Clean up StringReferenceValueComparator and DexFile.
Add a helper function to StringReference, so that the
StringReferenceValueComparator doesn't need to repeatedly
reference the same StringReference variable which led to
the bug 29602109.

Also clean up the DexFile, moving definitions of functions
that depend on the dex_file-inl.h from dex_file.h also to
the dex_file-inl.h.

Change-Id: I83e2aa67173c62b3432c26478e528bbb53b6d09f
Test: Run standard ART test suite on host and Nexus 5.
Bug: 29602109