ART: Do not add to JIT compile queue twice
Adding compile task to JIT queue creates a JNI global reference
which is released after processing this task. Number of JNI global
references is hardcoded and shared with JNI code. If JIT compilation
is slow the number of JNI global references may grow due to growth of
JIT compilation queue size.

We want to reduce the preasure by avoiding the adding new compile
task if it is already in the queue.

Change-Id: Iff3108e6ef0112c92bdc59ff4743c7969cfb9b8a
Signed-off-by: Vladimir <vladimir.a.ivanov@intel.com>
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>