[POC & RFC]: "simplify" our intrinsic recognizer.
This is just a proof of concept that we can use the modifiers
for storing the intrinsic kind. So lots of cleanup needed, but I
wanted to get a feel of what others think.

Note that this work was partly motivated by the fact that the inline
method analyzer leaks intrinsic tables, and even worst, might re-use
a table from one dex file to another unrelated dex file in the presence
of class unloading and the unlikely event of the dex files getting
the same address.

Change-Id: Ia653d2c72df13889dc85dd8c84997582c034ea4b