[WIP] Compiler switches for Built-in Bisection Search
This change implements two compiler switches to use in
Built-in Bisection Search: --optimize-up-to-method=<mthd_nr> and
--optimize-up-to-phase=<phase_nr>. When these switches are used
OptimizingCompiler enters a Bisection Mode where:
a) First <mthd_nr> - 1 methods are optimized fully.
b) Next method is optimized only up to <phase_nr>-th optimization
c) Following methods are not optimized at all
The logic is implemented with help of a new class: BisectionController.
This class is responsible for limiting the number of optimizations run.

Change-Id: I82f8913111d8ac5d4f922ea6f59701025b857efc