WIP: Use Thread.dispatchUncaughtException() rather than duplicating logic.
Previously, the runtime duplicated the logic in Thread's methods
dispatchUncaughtException() and getUncaughtExceptionHandler(). Since we're
modifying this logic to ensure that uncaught exceptions are logged, this
CL changes the runtime to call dispatchUncaughtException instead. This
also has the benefit that we can remove Thread.UncaughtExceptionHandler
from the list of well known classes.

Bug: 29624607
Change-Id: Id7161d841d64ff9d8f2a7ec1c8c0c5c911ff0b64