ART: Fix RegTypeCache for instance field declaring class
For bytecode instruction:

  iget(-*) ClassA->field:type

where the resolved field is actually declared in ClassB (ClassA is
a subclass of ClassB), MethodVerifier will create a RegType
representing ClassA but cache it under the descriptor of ClassA.

This is a bug but does not have any implications on correctness
because earlier resolution of ClassA (part of IGET handling) creates
another cache entry with the ClassA descriptor. Because the latter
entry is always imprecise (ClassB cannot be final because ClassA
extends it), either the earlier entry will be discovered or neither
of them will be a match.

Descriptor was replaced with the correct one and a DCHECK added when
creating a RegType to make sure the descriptor matches.

Bug: 30458218
Change-Id: I19e1bdd5dd79e5eac558122a06b9128d0a5c021f
Test: m test-art-host-run-test
Test: art/test/run-test --host 800-smali