Add ProcessBuilderTest#testRedirectInherit() to known failures
This test does fairly low level operations that previously exposed
other kernel issues. So far it has been observed failing 1/10 times
on an Android L kernel but never on CTS. Therefore, adding to list
of known failures. testRedirectErrorStream_outputAndErrorAreMerged
has not yet been observed to fail but is similar; it may need to be
added to the list if it fails in future.

Test: cts-tradefed run cts -c libcore.java.lang.ProcessBuilderTest

Change-Id: I5d67f67099e66cba4d75d9bf621609cd2d88d58a