[WIP] Change one read barrier bit to mark bit
Optimization to help slow path performance, when you slow path
on a reference. The GC sets the mark bit in the lock word of that
reference and avoids needing to go in the runtime the next time
the RB slow path is triggered.

TODO:
Handle mark stack overflow, prevent dirty mark bits in image, etc.

EAAC: 777ms -> 699ms (average 31 runs on N9)

Bug: 30162165

Change-Id: Iae0cacfae221e33151d3c0ab65338d1c822ab63d