Fixed bug in disassembly of roundss/roundsd
Rationale:
These instructions should be marked as load, so that, using
Intel syntax, destination appears at left hand side, as in
   roundss xmm1, xmm0
(and not roundss xmm0, xmm1). First I suspected a bug in the
encoding (hence the test) and even the register allocator, but
since the code behaved correctly, only disassembly was really wrong.

Test: disassembler_x86_test (but nothing for actual disassembly)

BUG=26327751

Change-Id: I060ef57f4d5a64cdc04b97ae8a799d1c0d22da05