Implement running user defined list of passes
This change introduces two new dex2oat switches --run-passes= and
--run-passes-zip=. These switches accept path to a text or zip file
respectivly. Compiler will only run optimization passes specified
in the file.

There is no verification implemented on the compiler side. It is user's
responsibility to provide a list of passes that leads to successful
generation of correct code.

We only take control of the optional optimizations. Codegen (builder),
and all passes required for register allocation will run unaffected
by this mechanism.

This change also enables usage of --compiled-methods switches and
disables the RestrictCompilationFiltersToImage flag. This is required
for built-in binary bug search.

Change-Id: Ic3694e53515fefcc5ce6f28d9371776b5afcbb4f