ART: Factor out native stack dumping, improve addr2line
Move DumpNativeStack and DumpKernelStack into a new file to keep
concerns separated. Clean up the file.

Change addr2line handling: use the tool in "server" mode of
operation, keeping a connection open for the same file. This
reduces the time of a dump of sleeping 001-HelloWorld from fifteen
to ten seconds.

Bug: 30351473
Test: m test-art-host
Test: manual host testing with a test that times out
Change-Id: Idbe424f85c566f5bb39d12644ce104fff54a2146